Add: JSDoc comments
