+++ src/components/event-add-button.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 /**
  * Класс кнопки добавления новых элементов
