+++ src/components/no-events.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 /**
  * Класс пустого списка точек маршрута
