+++ src/components/trip-sort.js
@@ -1,4 +1,4 @@
-import {createElement} from '../util.js';
+import {createElement} from '../utils/render.js';
 
 /**
  * Класс формы сортировки точек маршрута
