+++ src/controllers/board.js
@@ -6,20 +6,35 @@ import TripInfoComponent from '../components/trip-info.js';
 import TripDaysListComponent from '../components/days-list.js';
 
 import {generateTripDays, getTripCost, tripCards} from '../mocks/event.js';
-import {render, replace, RenderPosition} from '../util.js';
+import {render, replace, RenderPosition} from '../utils/render.js';
 
+/**
+* Рендеринг точки маршрута
+* @param {HTMLElement} container - Место вставки элемнета точки маршрута
+* @param {Object} eventCard - Данные для точки маршрута
+*/
 const renderTripEvent = (container, eventCard) => {
   const eventComponent = new EventComponent(eventCard);
   const eventEditComponent = new EventEditComponent(eventCard);
 
+  /**
+  * Замена карточку маршрута на форму редактирования точки марщрута
+  */
   const replaceCardToEdit = () => {
     replace(eventComponent, eventEditComponent);
   };
 
+  /**
+  * Замена формы редактирования точки марщрута на карточку маршрута
+  */
   const replaceEditToCard = () => {
     replace(eventEditComponent, eventComponent);
   };
 
+  /**
+  * Событие нажатия на Esc
+  * @param {evt} evt
+  */
   const onEscKeyDown = (evt) => {
     const isEscKey = evt.key === `Escape` || evt.key === `Esc`;
     if (isEscKey) {
@@ -28,15 +43,24 @@ const renderTripEvent = (container, eventCard) => {
     }
   };
 
+  /**
+  * Событие открытия формы редактирования при клике на стрелку
+  */
   eventComponent.setArrowBtnOpenHandler(() => {
     replaceCardToEdit();
     document.addEventListener(`keydown`, onEscKeyDown);
   });
 
+  /**
+  * Событие закрытия формы редактирования при клике на кнопку сброса
+  */
   eventEditComponent.setBtnResetHandler(() => {
     replaceEditToCard();
   });
 
+  /**
+  * Событие закрытия формы редактирования при клике на кнопку отправки
+  */
   eventEditComponent.setBtnSubmitHandler((evt) => {
     evt.preventDefault();
     replaceEditToCard();
@@ -45,6 +69,10 @@ const renderTripEvent = (container, eventCard) => {
   render(container, eventComponent.getElement(), RenderPosition.BEFOREEND);
 };
 
+/**
+* Рендеринг точки маршрута
+* @param {Object} tripCard - Данные для точки маршрута
+*/
 const renderTripEvents = (tripCard) => {
   const tripEventsList = document.querySelectorAll(`.trip-events__list`);
   tripEventsList.forEach((tripDay) => {
@@ -54,11 +82,18 @@ const renderTripEvents = (tripCard) => {
   });
 };
 
+/**
+ * Класс основной панели взаимодействия (инфрмация, сортировка, карточки
+ * либо приветственное собщение при остутствии карточек)
+ */
 export default class BoardController {
   constructor(container) {
     this._container = container;
   }
 
+  /**
+  * Рендеринг основной панели взаимодействия
+  */
   render() {
     const siteTripEventsElement = document.querySelector(`.trip-events`);
 
@@ -66,7 +101,6 @@ export default class BoardController {
       render(siteTripEventsElement, new NoEventsComponent().getElement(), RenderPosition.BEFOREEND);
     } else {
       const tripDays = generateTripDays(tripCards);
-      console.log(tripDays);
       const siteTripInfoElement = document.querySelector(`.trip-info`);
 
       render(siteTripEventsElement, new TripSortComponent().getElement(), RenderPosition.BEFOREEND);
